put m loop but iterate once

+++ crates/cubecl-linalg/src/matmul/components/stage/base.rs
@@ -173,4 +173,4 @@ pub enum StageBuffering {
 // TODO Support autotuning the best stage buffering
 //      However, from simple benchmarks on Maxime's computer (NVIDIA GeForce RTX 4060)
 //      Double seems to always be faster or comparable to simple.
-pub const STAGE_BUFFERING: StageBuffering = StageBuffering::Single;
+pub const STAGE_BUFFERING: StageBuffering = StageBuffering::Double;
